Rework the layout of the code a lot to make it """better"""
Lots of changes here in an attempt to organize the code better for
longer-term maintainability. Biggest change here is that things have
been shuffled into private cpp files, particularly implementation
details that are becoming problematic to include only in headers. This
has enabled some better encapsulation; this commit also includes some
extra namespacing and scoping for singleton access, etc. This commit
also introduces a new oscillator ui struct that's useful for managing
the UI thread's state.
