+++ CMakeLists.txt
@@ -21,8 +21,15 @@ file(GLOB_RECURSE
      "${CMAKE_SOURCE_DIR}"
      "${CMAKE_SOURCE_DIR}/include/*.h")
 
+# grab the private source files
+file(GLOB_RECURSE
+     _private_source_list
+     RELATIVE
+     "${CMAKE_SOURCE_DIR}"
+     "${CMAKE_SOURCE_DIR}/include/*.cpp")
+
 # gotta set WIN32 here or everything breaks
-add_executable(audiovisual WIN32 main.cpp ${_header_list})
+add_executable(audiovisual WIN32 main.cpp ${_private_source_list} ${_header_list})
 
 # properties of various sorts
 set_property(TARGET audiovisual PROPERTY CXX_STANDARD 20)
